actualizacion

+++ Solucion Control 02 - parcial/libreria/numPerfectos.java
@@ -1,7 +1,7 @@
 package libreria;
 
 public class numPerfectos {
-  public static boolean divisorPropio (int candidato, int numero) {
+  public static boolean divisorPropio (long candidato, long numero) {
     boolean esDivisorPropio;
     if ((candidato != numero) && ((numero % candidato) == 0)) {
       esDivisorPropio = true;
@@ -11,9 +11,9 @@ public class numPerfectos {
     return esDivisorPropio;
   }
   
-  public static int sumaDivisoresPropios (int numero) {
-    int total = 0;
-    for (int i=1; i<=numero/2; i++) {
+  public static long sumaDivisoresPropios (long numero) {
+    long total = 0;
+    for (long i=1; i<=numero/2; i++) {
       if (divisorPropio(i,numero)) {
         total += i;
       }
@@ -21,7 +21,7 @@ public class numPerfectos {
     return total;
   }
   
-  public static boolean numeroPerfect (int numero) {
+  public static boolean numeroPerfect (long numero) {
     boolean esNumeroPerfecto;
     if (numero == sumaDivisoresPropios(numero)) {
       esNumeroPerfecto = true;
